add logging support
