+++ backend/service/temporal/temporal.go
@@ -27,8 +27,8 @@ func New(cfg *anypb.Any, logger *zap.Logger, scope tally.Scope) (service.Service
 	tc, err := client.NewClient(client.Options{
 		HostPort:           fmt.Sprintf("%s:%d", config.Host, config.Port),
 		Namespace:          "",
-		Logger:             nil,
-		MetricsHandler:     &metricsHandler{Scope: scope},
+		Logger:             newTemporalLogger(logger),
+		MetricsHandler:     newMetricsHandler(scope),
 		Identity:           "",
 		DataConverter:      nil,
 		ContextPropagators: nil,
